src/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
SQLAlchemy>=1.4
marshmallow>=3.0
pymssql>=2.2
python-dotenv>=0.21
apispec>=6.0.0
apispec-webframeworks>=0.5.2
flask-swagger-ui>=4.11.1
PyJWT>=2.8
bcrypt>=4.1